[CTS] com.android.cts.browser.BrowserBenchTest#testOctane failed.
BZ: 156517

it is regression of the Google's Patch.
this is the commit: db22ec4ee014900988062d910bc810172a07df56
If the URL is already opened, switch to that tab, it will not reload this URL.
add this code to fix this issue.

Change-Id: I6a0e9626196e15a8f47f40eaa6137a039bacc1af
Category: aosp improvement
Domain: AOSP-Browser
Origin: internal
Upstream-Candidate: no, need rework
Signed-off-by: yijunx.zhu <yijunx.zhu@intel.com>